# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Ciph3r Product API",
        "description": "API RESTful para la gestión de productos, divisas y precios de productos.",
        "contact": {
            "email": "tu_email@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "Ciph3r API Server"
        }
    ],
    "paths": {
        "/currencies": {
            "get": {
                "tags": [
                    "Currencies"
                ],
                "summary": "Obtener lista de divisas",
                "description": "Retorna una lista de todas las divisas.",
                "operationId": "258b00af74ef78c350496a40fd13e890",
                "responses": {
                    "200": {
                        "description": "Lista de divisas obtenida exitosamente.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Currency"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Currencies"
                ],
                "summary": "Crear una nueva divisa",
                "description": "Crea una nueva divisa y la retorna.",
                "operationId": "a4b2e59d2eeca53302548306d848ee79",
                "requestBody": {
                    "description": "Datos de la divisa a crear",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreCurrencyRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Divisa creada exitosamente.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Currency"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": {
                                                "name": [
                                                    "The name field is required."
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/currencies/{id}": {
            "get": {
                "tags": [
                    "Currencies"
                ],
                "summary": "Obtener divisa por ID",
                "description": "Retorna una divisa específica por su ID.",
                "operationId": "1206b2c5eab8182e5d090a132aa7daa8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la divisa",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Divisa obtenida exitosamente.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Currency"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Divisa no encontrada.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Resource not found."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Currencies"
                ],
                "summary": "Actualizar divisa",
                "description": "Actualiza una divisa existente por su ID y la retorna.",
                "operationId": "e661ab392e0bc735216d7d595db42999",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la divisa a actualizar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Datos de la divisa a actualizar",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateCurrencyRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Divisa actualizada exitosamente.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Currency"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Divisa no encontrada.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Resource not found."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación."
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Currencies"
                ],
                "summary": "Eliminar divisa",
                "description": "Elimina una divisa por su ID.",
                "operationId": "375a808f3b36a5226492c20c35b46103",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la divisa a eliminar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Divisa eliminada exitosamente (No Content)."
                    },
                    "404": {
                        "description": "Divisa no encontrada.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Resource not found."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Currency": {
                "title": "Currency",
                "description": "Modelo de una divisa",
                "properties": {
                    "id": {
                        "description": "Identificador único de la divisa",
                        "type": "integer",
                        "format": "int64"
                    },
                    "name": {
                        "description": "Nombre de la divisa (ej. US Dollar)",
                        "type": "string"
                    },
                    "symbol": {
                        "description": "Símbolo de la divisa (ej. USD)",
                        "type": "string"
                    },
                    "exchange_rate": {
                        "description": "Tasa de cambio de la divisa respecto a una base (ej. 1.0 para USD)",
                        "type": "number",
                        "format": "float"
                    },
                    "created_at": {
                        "description": "Fecha y hora de creación",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Fecha y hora de última actualización",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "StoreCurrencyRequest": {
                "title": "Store Currency Request",
                "description": "Datos para crear una nueva divisa",
                "required": [
                    "name",
                    "symbol",
                    "exchange_rate"
                ],
                "properties": {
                    "name": {
                        "description": "Nombre de la divisa (ej. US Dollar)",
                        "type": "string",
                        "example": "US Dollar"
                    },
                    "symbol": {
                        "description": "Símbolo de la divisa (ej. USD)",
                        "type": "string",
                        "example": "USD"
                    },
                    "exchange_rate": {
                        "description": "Tasa de cambio de la divisa (ej. 1.0)",
                        "type": "number",
                        "format": "float",
                        "example": 1
                    }
                },
                "type": "object"
            },
            "UpdateCurrencyRequest": {
                "title": "Update Currency Request",
                "description": "Datos para actualizar una divisa",
                "required": [
                    "name",
                    "symbol",
                    "exchange_rate"
                ],
                "properties": {
                    "name": {
                        "description": "Nombre de la divisa (ej. US Dollar)",
                        "type": "string",
                        "example": "US Dollar"
                    },
                    "symbol": {
                        "description": "Símbolo de la divisa (ej. USD)",
                        "type": "string",
                        "example": "USD"
                    },
                    "exchange_rate": {
                        "description": "Tasa de cambio de la divisa (ej. 1.0)",
                        "type": "number",
                        "format": "float",
                        "example": 1
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Autenticación con Laravel Sanctum. Genera un token usando /api/auth/login y envíalo en el encabezado 'Authorization: Bearer {token}'.",
                "name": "Sanctum Token based Auth",
                "in": "header",
                "bearerFormat": "Sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}